refatorar código e ajustar função de filtragem
