+++ index.html
@@ -56,6 +56,6 @@
       <h3>Seu relatório será gerado aqui.</h3>
     </div>
   </div>
-  <script src="index.js"></script>
+  <script src="./js/index.js" type="module"></script>
 </body>
 </html>
